Added different preconditioning algorithms, updated tests
As discussed, three different preconditioning algorithms should be supported:
Gauss-Seidel, ILU0 and SPAI0.

The implementation for including these options is really ugly - the reason
is that the AMGCL library does not contain abstractions to build on, and
figuring out an elegant way would have taken very long.

Tests were updated to check new cases, with the same matrix generation as before.